layout/xul/test/test_bug398982-2.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        xmlns:html="http://www.w3.org/1999/xhtml"
        title="Test for Bug 398982">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=398982
-->
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

<popupgroup style="position: absolute; ">
<tooltip type="zzz">
<treecols/>
</tooltip>

<script xmlns="http://www.w3.org/1999/xhtml" class="testbody" type="application/javascript">
<![CDATA[
function doe() {
  document.getElementsByTagName('popupgroup')[0].removeAttribute('style');
  is(0, 0, "Test is successful if we get here without crashing");
  SimpleTest.finish();
}
function do_test() {
  setTimeout(doe, 200);
}
SimpleTest.waitForExplicitFinish();
addLoadEvent(do_test);
]]>
</script>
</popupgroup>
<html:body></html:body> <!-- XXX SimpleTest.showReport() requires a html:body -->
</window>
